- Added character selection. Respawn will now wait for a valid character selection. - Fixed the spring triggering when hitting it from below. - Optimized updating level blocks through an RPC call, instead of serializing the entire level each time a block is changed. - Improved bullet collision. Collision is now rechecked every 4 pixels. - Tweaked bullet speed.

+++ users/marcel/netarena/host.cpp
@@ -140,7 +140,7 @@ void Host::tickPlay(float dt)
 
 	if (time >= m_nextPickupSpawnTime)
 	{
-		if (m_pickups.size() < g_pickupMax)
+		if ((int)m_pickups.size() < g_pickupMax)
 		{
 			int x;
 			int y;
